+++ webknossos/Changelog.md
@@ -15,6 +15,7 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 ### Breaking Changes
 
 ### Added
+- Added method download_mesh() to the Annotation class to allow download of .stl files. [#1307](https://github.com/scalableminds/webknossos-libs/pull/1307)
 
 ### Changed